Initial Tests

+++ src/controllers/campaigns.ts
@@ -1,35 +1,47 @@
 import { Request, Response } from "express";
 import { CampaignService } from "../services/campaigns";
 import { Campaign, ICreateCampaign } from "../interfaces";
+import { StatusCodes } from "http-status-codes";
+import { isValidId } from "../utils/validation";
 
 async function getAllCampaigns(req: Request, res: Response) {
   try {
     const campaigns = await CampaignService.getAllCampaigns();
     return res.status(200).json(campaigns);
   } catch (error) {
-    res.status(401).json("Cannot access database");
+    res.status(500).json("Cannot access database");
   }
 }
 
 async function getCampaignById(req: Request, res: Response) {
   try {
     const campaignId = parseInt(req.params["id"]);
+    if (!isValidId(campaignId)) {
+      return res
+        .status(StatusCodes.BAD_REQUEST)
+        .json("Requires valid campaignId");
+    }
     const campaign = await CampaignService.getCampaignById(campaignId);
     return res.status(200).json(campaign);
   } catch (error) {
-    res.status(401).json("Cannot find campaign id");
+    res.status(500).json("Cannot find campaign id");
   }
 }
 
 async function getCampaignsByInfluencers(req: Request, res: Response) {
   try {
     const influencerId = parseInt(req.params["id"]);
+    if (!isValidId(influencerId)) {
+      return res
+        .status(StatusCodes.BAD_REQUEST)
+        .json("Requires valid influencerId");
+    }
     const campaigns = await CampaignService.getCampaignsByInfluencer(
       influencerId
     );
     return res.status(200).json(campaigns);
   } catch (error) {
-    res.status(401).json("Cannot find campaign id");
+    res.status(500).json("Cannot find campaign id");
   }
 }
 
@@ -57,6 +69,11 @@ async function createCampaign(req: Request, res: Response) {
 
 async function deleteCampaignById(req: Request, res: Response) {
   const { campaignId: campaignId } = req.body;
+  if (!isValidId(campaignId)) {
+    return res
+      .status(StatusCodes.BAD_REQUEST)
+      .json("Requires valid campaignId");
+  }
   const deletedCampaign = await CampaignService.deleteCampaignById(campaignId);
   if (!deletedCampaign) {
     return res.status(500).json("Cannot delete id");
