+++ src/controllers/clients.ts
@@ -1,33 +1,43 @@
 import { Request, Response } from "express";
 import { ClientService } from "../services/clients";
 import { Client, ICreateClient } from "../interfaces";
+import { isValidId } from "../utils/validation";
+import { StatusCodes } from "http-status-codes";
 
 async function getAllClients(req: Request, res: Response) {
   try {
     const clients = await ClientService.getAllClients();
     return res.status(200).json(clients);
   } catch (error) {
-    res.status(401).json("Cannot access database");
+    res.status(500).json("Cannot access database");
   }
 }
 
 async function getClientsByUserId(req: Request, res: Response) {
   try {
     const userId = parseInt(req.params["id"]);
+    if (!isValidId(userId)) {
+      return res.status(StatusCodes.BAD_REQUEST).json("Requires valid userId");
+    }
     const clients = await ClientService.getClientsByUserId(userId);
     return res.status(200).json(clients);
   } catch (error) {
-    res.status(401).json("Cannot access database");
+    res.status(500).json("Cannot access database");
   }
 }
 
 async function getClientById(req: Request, res: Response) {
   try {
     const clientId = parseInt(req.params["id"]);
+    if (!isValidId(clientId)) {
+      return res
+        .status(StatusCodes.BAD_REQUEST)
+        .json("Requires valid clientId");
+    }
     const client = await ClientService.getClientById(clientId);
     return res.status(200).json(client);
   } catch (error) {
-    res.status(401).json("Cannot find client id");
+    res.status(500).json("Cannot find client id");
   }
 }
 
@@ -55,6 +65,9 @@ async function createClient(req: Request, res: Response) {
 
 async function deleteClientById(req: Request, res: Response) {
   const { clientId: clientId } = req.body;
+  if (!isValidId(clientId)) {
+    return res.status(StatusCodes.BAD_REQUEST).json("Requires valid clientId");
+  }
   const deletedClient = await ClientService.deleteClientById(clientId);
   if (!deletedClient) {
     return res.status(500).json("Cannot delete id");
